line-provider/pyproject.toml
[tool.poetry]
name = "line-provider"
version = "0.1.0"
description = "Service of information about events on which you can bet."
authors = ["Ilya Viarbilo <verbiloilyaa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.112.2"}
pydantic-settings = "^2.4.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.21.6"}
aerich = "^0.7.2"
aio-pika = "^9.4.3"



[tool.aerich]
tortoise_orm = "line_provider.db.config.TORTOISE_CONFIG"
location = "./line_provider/db"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
